% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl.R
\name{.compute_tail_factor}
\alias{.compute_tail_factor}
\title{Compute log-linear tail factor from selected ATA factors}
\usage{
.compute_tail_factor(selected, tail)
}
\description{
Compute log-linear tail factor from selected ATA factors
}
\keyword{internal}
