% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed.R
\name{.extrapolate_sigma_ed}
\alias{.extrapolate_sigma_ed}
\title{Extrapolate missing sigma values for ED links}
\usage{
.extrapolate_sigma_ed(x, method = c("min_last2", "locf", "loglinear"))
}
\description{
Extrapolate missing sigma values for ED links
}
\keyword{internal}
