% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experience.R
\name{is_experience}
\alias{is_experience}
\title{Check whether an object is an \code{Experience}}
\usage{
is_experience(x)
}
\arguments{
\item{x}{An object.}
}
\value{
Logical scalar.
}
\description{
Check whether an object is an \code{Experience}
}
