% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ata-vis.R
\name{plot.ata_fit}
\alias{plot.ata_fit}
\title{Plot an ata fit}
\usage{
\method{plot}{ata_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"ata_fit"}.}

\item{...}{Arguments passed to \code{\link[=plot.ata]{plot.ata()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualise an object of class \code{"ata_fit"} by delegating to \code{\link[=plot.ata]{plot.ata()}}
on the underlying \code{ata} data stored in \code{x$ata}.
}
\seealso{
\code{\link[=plot.ata]{plot.ata()}}, \code{\link[=fit_ata]{fit_ata()}}
}
