% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ata-vis.R
\name{plot.ATAFit}
\alias{plot.ATAFit}
\title{Plot an ata fit}
\usage{
\method{plot}{ATAFit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"ATAFit"}.}

\item{...}{Arguments passed to \code{\link[=plot.ATA]{plot.ATA()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualise an object of class \code{"ATAFit"} by delegating to \code{\link[=plot.ATA]{plot.ATA()}}
on the underlying \code{ATA} data stored in \code{x$ata}.
}
\seealso{
\code{\link[=plot.ATA]{plot.ATA()}}, \code{\link[=fit_ata]{fit_ata()}}
}
