% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed-vis.R
\name{plot.ed}
\alias{plot.ed}
\title{Plot ED intensity diagnostics}
\usage{
\method{plot}{ed}(
  x,
  type = c("summary", "box", "point"),
  alpha = 1,
  scales = c("fixed", "free", "free_x", "free_y"),
  nrow = NULL,
  ncol = NULL,
  theme = c("view", "save", "shiny"),
  ...
)
}
\arguments{
\item{x}{An object of class \code{"ed"}.}

\item{type}{One of \code{"summary"}, \code{"box"}, or \code{"point"}.}

\item{alpha}{Numeric scalar. Default is \code{1}.}

\item{scales}{Facet scale argument.}

\item{nrow, ncol}{Facet dimensions.}

\item{theme}{Theme string.}

\item{...}{Additional arguments passed to \code{\link[=.switch_theme]{.switch_theme()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualise diagnostic summaries from an \code{"ed"} object. Internally
calls \code{\link[=summary_ed]{summary_ed()}}.
}
