% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed-vis.R
\name{plot.ed_fit}
\alias{plot.ed_fit}
\title{Plot an ED fit}
\usage{
\method{plot}{ed_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"ed_fit"}.}

\item{...}{Arguments passed to \code{\link[=plot.ed]{plot.ed()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualise an object of class \code{"ed_fit"} by delegating to \code{\link[=plot.ed]{plot.ed()}}
on the underlying \code{ed} data stored in \code{x$ed}.
}
\seealso{
\code{\link[=plot.ed]{plot.ed()}}, \code{\link[=fit_ed]{fit_ed()}}
}
