% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed-vis.R
\name{plot.EDFit}
\alias{plot.EDFit}
\title{Plot an ED fit}
\usage{
\method{plot}{EDFit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"EDFit"}.}

\item{...}{Arguments passed to \code{\link[=plot.ED]{plot.ED()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualise an object of class \code{"EDFit"} by delegating to \code{\link[=plot.ED]{plot.ED()}}
on the underlying \code{ED} data stored in \code{x$ed}.
}
\seealso{
\code{\link[=plot.ED]{plot.ED()}}, \code{\link[=fit_ed]{fit_ed()}}
}
