% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ata-vis.R
\name{plot_triangle.ata_fit}
\alias{plot_triangle.ata_fit}
\title{Triangle heatmap for an ata fit}
\usage{
\method{plot_triangle}{ata_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"ata_fit"}.}

\item{...}{Arguments passed to \code{\link[=plot_triangle.ata]{plot_triangle.ata()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualise an object of class \code{"ata_fit"} as a triangle-style heatmap
by delegating to \code{\link[=plot_triangle]{plot_triangle()}} on the underlying \code{ata} data
stored in \code{x$ata}.
}
\seealso{
\code{\link[=plot_triangle.ata]{plot_triangle.ata()}}, \code{\link[=fit_ata]{fit_ata()}}
}
