% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed-vis.R
\name{plot_triangle.ed_fit}
\alias{plot_triangle.ed_fit}
\title{Triangle heatmap for an ED fit}
\usage{
\method{plot_triangle}{ed_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"ed_fit"}.}

\item{...}{Arguments passed to \code{\link[=plot_triangle.ed]{plot_triangle.ed()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualise an object of class \code{"ed_fit"} as a triangle-style heatmap
by delegating to \code{\link[=plot_triangle.ed]{plot_triangle.ed()}} on the underlying \code{ed} data
stored in \code{x$ed}.
}
\seealso{
\code{\link[=plot_triangle.ed]{plot_triangle.ed()}}, \code{\link[=fit_ed]{fit_ed()}}
}
