% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed-vis.R
\name{plot_triangle.EDFit}
\alias{plot_triangle.EDFit}
\title{Triangle heatmap for an ED fit}
\usage{
\method{plot_triangle}{EDFit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"EDFit"}.}

\item{...}{Arguments passed to \code{\link[=plot_triangle.ED]{plot_triangle.ED()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualise an object of class \code{"EDFit"} as a triangle-style heatmap
by delegating to \code{\link[=plot_triangle.ED]{plot_triangle.ED()}} on the underlying \code{ED} data
stored in \code{x$ed}.
}
\seealso{
\code{\link[=plot_triangle.ED]{plot_triangle.ED()}}, \code{\link[=fit_ed]{fit_ed()}}
}
