% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle-vis.R
\name{plot_triangle}
\alias{plot_triangle}
\title{Triangle plot generic}
\usage{
plot_triangle(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A plot object.
}
\description{
Generic function for triangle-style visualisations.
}
