% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ata.R
\name{print.ata_fit}
\alias{print.ata_fit}
\title{Print an \code{ata_fit} object}
\usage{
\method{print}{ata_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"ata_fit"}.}

\item{...}{Unused.}
}
\description{
Print an \code{ata_fit} object
}
