% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ata.R
\name{print.ATAFit}
\alias{print.ATAFit}
\title{Print an \code{ATAFit} object}
\usage{
\method{print}{ATAFit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"ATAFit"}.}

\item{...}{Unused.}
}
\description{
Print an \code{ATAFit} object
}
