% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl.R
\name{print.cl_fit}
\alias{print.cl_fit}
\title{Print a \code{cl_fit} object}
\usage{
\method{print}{cl_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"cl_fit"}.}

\item{...}{Unused.}
}
\description{
Print a \code{cl_fit} object
}
