% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl.R
\name{print.CLFit}
\alias{print.CLFit}
\title{Print a \code{CLFit} object}
\usage{
\method{print}{CLFit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"CLFit"}.}

\item{...}{Unused.}
}
\description{
Print a \code{CLFit} object
}
