% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed.R
\name{print.ed_fit}
\alias{print.ed_fit}
\title{Print an \code{ed_fit} object}
\usage{
\method{print}{ed_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"ed_fit"}.}

\item{...}{Unused.}
}
\description{
Print an \code{ed_fit} object
}
