% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ed.R
\name{print.EDFit}
\alias{print.EDFit}
\title{Print an \code{EDFit} object}
\usage{
\method{print}{EDFit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"EDFit"}.}

\item{...}{Unused.}
}
\description{
Print an \code{EDFit} object
}
