% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr.R
\name{print.lr_fit}
\alias{print.lr_fit}
\title{Print an \code{lr_fit} object}
\usage{
\method{print}{lr_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"lr_fit"}.}

\item{...}{Unused.}
}
\description{
Print an \code{lr_fit} object
}
