% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr.R
\name{print.LRFit}
\alias{print.LRFit}
\title{Print an \code{LRFit} object}
\usage{
\method{print}{LRFit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"LRFit"}.}

\item{...}{Unused.}
}
\description{
Print an \code{LRFit} object
}
