% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl.R
\name{summary.cl_fit}
\alias{summary.cl_fit}
\title{Summary method for \code{cl_fit}}
\usage{
\method{summary}{cl_fit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"cl_fit"}.}

\item{...}{Unused.}
}
\value{
A \code{data.table} with one row per cohort.
}
\description{
Summary method for \code{cl_fit}
}
