% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl.R
\name{summary.CLFit}
\alias{summary.CLFit}
\title{Summary method for \code{CLFit}}
\usage{
\method{summary}{CLFit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"CLFit"}.}

\item{...}{Unused.}
}
\value{
A \code{data.table} with one row per cohort.
}
\description{
Summary method for \code{CLFit}
}
