% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr.R
\name{summary.lr_fit}
\alias{summary.lr_fit}
\title{Summary method for \code{lr_fit}}
\usage{
\method{summary}{lr_fit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"lr_fit"}.}

\item{...}{Unused.}
}
\value{
A \code{data.table} with one row per cohort.
}
\description{
Summary method for \code{lr_fit}
}
