% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr.R
\name{summary.LRFit}
\alias{summary.LRFit}
\title{Summary method for \code{LRFit}}
\usage{
\method{summary}{LRFit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"LRFit"}.}

\item{...}{Unused.}
}
\value{
A \code{data.table} with one row per cohort.
}
\description{
Summary method for \code{LRFit}
}
