% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lossratio-package.R
\docType{package}
\name{lossratio-package}
\alias{lossratio}
\alias{lossratio-package}
\title{The chain-ladder technique is only accurate when patterns of loss development
in the past can be assumed to continue in the future}
\description{
Applying chain ladder methods to predict the loss ratio for Korean health care products, not for loss reserving.
}
\author{
\strong{Maintainer}: Seokhoon Joo \email{seokhoonj@gmail.com}

}
\keyword{internal}
